requirements.txt
-i https://pypi.org/simple
absl-py==0.7.1
aniso8601==7.0.0
astor==0.8.0
cassandra-driver==3.18.0
cffi==1.12.3 ; sys_platform == 'win32' and platform_python_implementation == 'CPython'
click==7.0
flask-restful==0.3.7
flask==1.0.3
gast==0.2.2
gevent==1.4.0
greenlet==0.4.15 ; platform_python_implementation == 'CPython'
grpcio==1.21.1
h5py==2.9.0
itsdangerous==1.1.0
jinja2==2.10.1
keras-applications==1.0.8
keras-preprocessing==1.1.0
keras==2.2.4
markdown==3.1.1
markupsafe==1.1.1
mock==3.0.5
numpy==1.16.4
protobuf==3.8.0
pycparser==2.19
pytz==2019.1
pyyaml==5.1.1
scipy==1.3.0
six==1.12.0
tensorboard==1.13.1
tensorflow-estimator==1.13.0
tensorflow==1.13.1
termcolor==1.1.0
waitress==1.3.0
werkzeug==0.15.4
wheel==0.33.4 ; python_version >= '3'
